function call change and new ordering
Changing how the function is called in PowerShell as it makes it easier to CTRL+F and find where logging occurs on a per function level. i.e. New-SMEXCOEvent -Source "General" can be done unique to all of the function calls but New-SMEXCOEvent -ID 0 would span several different types of functions as this functionality grows.

Also re-ordering the event numbering to log a verbose event to show the connection to Exchange was succesful