more logging for functions (#211)
* function call change and new ordering

Changing how the function is called in PowerShell as it makes it easier to CTRL+F and find where logging occurs on a per function level. i.e. New-SMEXCOEvent -Source "General" can be done unique to all of the function calls but New-SMEXCOEvent -ID 0 would span several different types of functions as this functionality grows.

Also re-ordering the event numbering to log a verbose event to show the connection to Exchange was succesful

* Schedule-WorkItem and Get-SCSMUserByEmailAddress

adding logging options for setting/nulling Scheduled Work Items with appointments/meetings as well as details around Get-SCSMUserByEmailAddress

* Create-UserInCMDB

Opting to make this part of regular Informational events as opposed to requiring Verbose logging. Introducing a variable to hold the projection -passthru result to test for existence on the optional logging step

* Send-EmailFromWorkflowAccount and Test-KeywordsFoundInMessage

adding logging for when an email is sent from the workflow account and if a match is found when attempting to match keywords to drive IR/SR

* event call order

missed re-ordering one of the new-smexcoevent calls

* evening out event parameters

While they aren't getting used just yet, the amount of incoming parameters did not equal the amount that could be set

* logging sources

adding more logging sources to keep parity with inbound parameters