portal workflow integration (#214)
* cireson portal token call

updating the call to retrieve an auth token from the Cireson portal depending on if workflows are being used or not.

* portal secure reference

if the secure reference can't be set, it must have been left blank in the UI. This forces the selection of the first secure reference to be set to the mp so we can ultimately set this as a workflow parameter.

* workflow rule update action

When the form is saved, the rule is updated with 3 new parameters for use with optional cireson portal integration. This approach dynamically the Workflow Parameters and then inserts those Workflow Parameters into the Workflow to be set with the WriteActionCollection

* workflow rule create action

When the rule doesn't exist, the rule is created with 6 parameters for use with optional cireson portal integration. This approach dynamically builds the Workflow Parameters and then inserts those Workflow Parameters into the Workflow to be set with the WriteActionCollection

* portal params

adding new parameters to for authenticating with the cireson portal

* token call change

updating call to retrieve portal tokens

* variable naming

improving variable names