impersonation for workflow
With Impersonation, the email address used for $workflowEmailAddress is the SCSM Inbox that receives mail but the account that authenticates to it (e.g. the selected Run as Account) is _not_ the actual AD User tied to the address.

Thus if the workflow is using Impersonation, further define the $exchangeService object with the ImpersonatedUserId.