Exchange Impersonation (#215)
* impersonation bool

Introducing a new settings value for deciding whether or not Exchange Impersonation is used

* impersonation checkbox

adding a checkbox to control whether or not impersonation is used when authenticating to exchange

* bindings for Impersonation

adding variables and bindings to the UI to save back to the Admin Settings class and load the value

* impersonation for workflow

With Impersonation, the email address used for $workflowEmailAddress is the SCSM Inbox that receives mail but the account that authenticates to it (e.g. the selected Run as Account) is _not_ the actual AD User tied to the address.

Thus if the workflow is using Impersonation, further define the $exchangeService object with the ImpersonatedUserId.