digital signature enhancements
introducing logic that checks the validity of the digital signature. If the signature can't be validated and custom actions are enabled, a new custom action can be triggered. If the signature can be validated and the new PowerShell keyword exists - trigger a custom action. $ignoreValidSig and $pwshKeyword will eventually be defined in the Settings MP as a bool and string respectively.