Digital Signature Custom Actions (#225)
* digital signature enhancements

introducing logic that checks the validity of the digital signature. If the signature can't be validated and custom actions are enabled, a new custom action can be triggered. If the signature can be validated and the new PowerShell keyword exists - trigger a custom action. $ignoreValidSig and $pwshKeyword will eventually be defined in the Settings MP as a bool and string respectively.

* property - Ignore Invalid Digital Signature

adding a new bool property to the settings mp that allows the connector skip or process digitally signed email with an invalid digital signature

* bind - Ignore Invalid Digital Signature

Variables and bindings for the UI to save the checkbox value (true/false) to the property

* UI - Ignore Invalid Digital Signature

Adding checkbox to the UI that allows control of processing or ignoring digitally signed emails whose signature is not valid

* pwsh - Ignore Invalid Digital Signature

Updating PowerShell logic to pull the stored true/false value from the MP to process accordingly

* property - pwsh keyword

adding a new string keyword to the settings mp that allows the connector to invoke custom actions if the digital signature can be verified

* bind - pwsh keyword

Variables and bindings for the UI to save the new powershell keyword

* UI - pwsh keyword

Adding textbox to the UI that allows you to set the powershell (custom action) keyword to be used when a digital signature can be verified

* pwsh - keyword for powershell

Updating PowerShell logic to pull/reference the PowerShell keyword value if it's featured in the subject, custom actions are enabled, and the email's digital signature can be verified

* digital signature events

Updating custom events with new optional functions that can be invoked

* logging options

Introducing logging for digital signature validity

* phrasing

The grammar used here is misleading and now better aligns with the choices above it. Also indenting the checkbox here.

* new icon

adding graphic for readme

* pwsh keyword

adding verbiage for powershell keyword functionality