Regional decimal formatting (#227)
* dropping regex validation

Since decimal points can vary between regions, either the regex needs to be expanded or solved in the code behind.

* globalization reference

adding System.Globalization to provide region specific settings as they pertain to the Settings UI

* culture and region declarations

adding the current region and allowed number formats as variables to use during various parsing operations in the Settings UI during the save/commit action i.e. when the "OK" button is pushed

* culture - Min File Size

Updating the logic used to commit the Minimum File Size decimal value. First, an attempt is made to parse the value entered per the current user's Windows Region. Assuming it's valid, the value becomes culture invariant so the decimal delimiter stored in the MP/database is a period.

* culture - Min ACS to Create SR

Updating the logic used to commit the Minimum ACS score that creates an SR

* culture - Announcement Expiration

Updating the logic used to commit the time in hours that Announcements expire

* culture - Azure Machine Learning

Updating the logic used to commit the min confidence percent used by AML