v3.1 (#226)
* signing/encryption logic (#219)

* signing/encryption logic

updating the logic for how messages that are digitally signed and/or encrypted are processed. Message body now can be successfully extracted from the mime message

* Mimekit v2.9.2

Upgrading the version of Mimekit used by the connector

* property change

updating property used for To and Cc

* mimekit logging (#220)

Adding logging event for MimeKit/cryptography

* logging change (#221)

didn't remove the original decryption call that existed outside of logging

* Mimekit & BouncyCastle (#222)

* mimekit + bouncycastle

introduction of bouncycastle crypto for C# v1.8.5.50 as MimeKit requires/builds with it

* dll load for BouncyCastle

loads BouncyCastle crypto dll based on the directory that MimeKit is located in

* Digital Signature Custom Actions (#225)

* digital signature enhancements

introducing logic that checks the validity of the digital signature. If the signature can't be validated and custom actions are enabled, a new custom action can be triggered. If the signature can be validated and the new PowerShell keyword exists - trigger a custom action. $ignoreValidSig and $pwshKeyword will eventually be defined in the Settings MP as a bool and string respectively.

* property - Ignore Invalid Digital Signature

adding a new bool property to the settings mp that allows the connector skip or process digitally signed email with an invalid digital signature

* bind - Ignore Invalid Digital Signature

Variables and bindings for the UI to save the checkbox value (true/false) to the property

* UI - Ignore Invalid Digital Signature

Adding checkbox to the UI that allows control of processing or ignoring digitally signed emails whose signature is not valid

* pwsh - Ignore Invalid Digital Signature

Updating PowerShell logic to pull the stored true/false value from the MP to process accordingly

* property - pwsh keyword

adding a new string keyword to the settings mp that allows the connector to invoke custom actions if the digital signature can be verified

* bind - pwsh keyword

Variables and bindings for the UI to save the new powershell keyword

* UI - pwsh keyword

Adding textbox to the UI that allows you to set the powershell (custom action) keyword to be used when a digital signature can be verified

* pwsh - keyword for powershell

Updating PowerShell logic to pull/reference the PowerShell keyword value if it's featured in the subject, custom actions are enabled, and the email's digital signature can be verified

* digital signature events

Updating custom events with new optional functions that can be invoked

* logging options

Introducing logging for digital signature validity

* phrasing

The grammar used here is misleading and now better aligns with the choices above it. Also indenting the checkbox here.

* new icon

adding graphic for readme

* pwsh keyword

adding verbiage for powershell keyword functionality

* Regional decimal formatting (#227)

* dropping regex validation

Since decimal points can vary between regions, either the regex needs to be expanded or solved in the code behind.

* globalization reference

adding System.Globalization to provide region specific settings as they pertain to the Settings UI

* culture and region declarations

adding the current region and allowed number formats as variables to use during various parsing operations in the Settings UI during the save/commit action i.e. when the "OK" button is pushed

* culture - Min File Size

Updating the logic used to commit the Minimum File Size decimal value. First, an attempt is made to parse the value entered per the current user's Windows Region. Assuming it's valid, the value becomes culture invariant so the decimal delimiter stored in the MP/database is a period.

* culture - Min ACS to Create SR

Updating the logic used to commit the Minimum ACS score that creates an SR

* culture - Announcement Expiration

Updating the logic used to commit the time in hours that Announcements expire

* culture - Azure Machine Learning

Updating the logic used to commit the min confidence percent used by AML

* inline notes

updating inline notes in prep for #226

* readme graphic transparent background (#230)

updating readme graphic to have use a transparent background

* incrementing build number (#232)

upping version numbers for files