GitHub Actions: Build and Release (#240)
* build and release

GitHub Action that compiles the connector's C# management pack and other dependencies. Then creates a draft release with the attached zip. Takes three inputs - branch name, version, and if the Release to be drafted should be marked as release/pre-release

* github dependencies

new dependencies required for GitHub Action to compile the sln natively on GitHub

* github dependencies

VSAE schemas required for building through GitHub Actions

* build and release (directory change)

GitHub Action that compiles the connector's C# management pack and other dependencies. Then creates a draft release with the attached zip. Takes three inputs - branch name, version, and if the Release to be drafted should be marked as release/pre-release

* moving directories

placed in /workflows

* updating SLN for GitHub

Fixing: The project "SMLets.Exchange.Connector" is not selected for building in solution configuration "Debug|Any CPU".