delete message after processing bool
Introducing a new settings value for handling what happens to mail after it's been processed. This is to mirror stock connector functionality so that the message is either kept in the inbox as read so it's ignored on subsequent runs of the connector _or_ move to mailbox's deleted items as is current functionality.