delete message after processing (#260)
* delete message after processing bool

Introducing a new settings value for handling what happens to mail after it's been processed. This is to mirror stock connector functionality so that the message is either kept in the inbox as read so it's ignored on subsequent runs of the connector _or_ move to mailbox's deleted items as is current functionality.

* move to deleted items checkbox

adding a checkbox in the UI to control whether or not the message is moved to the inbox's Deleted Items (i.e. Trash) after processing

* bindings - move to deleted items after processing

adding variables and bindings to the UI to save back to the Admin Settings class and load the value for "Move to Deleted Items After Processing"

* move to deleted to items for pwsh

If the UI is set to true/checked off. Then messages will be moved to the Deleted Items/Trash of the Inbox. If it's false/unchecked, messages will remain in the Inbox but marked as Read so as to avoid subsequent processing.