OAuth error handling (#278)
When connecting to 365 it's possible that the SCSM Run As Account is entered as simply a username and the domain selected VS. entering the username as an email address thereby preventing the selection of the domain. This distinction is also made in the stock Exchange Connector documentation as follows "Ensure that the email address of the account that you enter is in down-level logon name format (for example, DOMAIN\UserName)". In either case, both formats ensure the use of "domain.tld" which grants a successful connection.