Extending logging (#273)
* Get-SCSMWorkItemParent logging

Updating the Get-SCSMWorkItemParent function with New-SMExcoEvent functions instead of write-verbose

* Get-AzureEmailLanguage logging

Adding logging for when the language of the email is identified by Azure Cognitive Services

* Get-TierMembers logging

Adding logging for when the users of a Support Group (enum) are requested

* Get-AssignedToWorkItemVolume logging

Introducing logging when a user's active assigned work count is requested. Given the potential size of a support group, this requires verbose logging as it could potentially generate a lot of events.