v3.3 (#275)
* declaring dependency (#274)

Declaring MimeKit as hosted on Github as a dependency for the project

* technet links (#271)

updating links to technet articles that are now hosted on the System Center blog

* codeql badge (#272)

adding CodeQL security scanning status to readme

* Additional Attachment handling (#277)

* Additional Attachment handling

Updating the Attach-FileToWorkItem function to support File and Item Attachments as well if they are Encrypted and or Signed

* signed/encrypted attachments

Extending digitally signed/encrypted messages to support cases wherein they have File/Item Attachments

* improved RE: logic (#270)

extending use cases for Merge Replies functionality

* OAuth error handling (#278)

When connecting to 365 it's possible that the SCSM Run As Account is entered as simply a username and the domain selected VS. entering the username as an email address thereby preventing the selection of the domain. This distinction is also made in the stock Exchange Connector documentation as follows "Ensure that the email address of the account that you enter is in down-level logon name format (for example, DOMAIN\UserName)". In either case, both formats ensure the use of "domain.tld" which grants a successful connection.

* History UI (#279)

* load settings history

when the UI is loaded, history is retrieved and stored in memory

* History Form in Build

When the project is built, include the History form

* Form and Code Behind

History Form and History Form code behind placeholders. Will compile with build, but currently do not render the History in the UI.

* History Form in UI

Including the History Form as a Wizard Step so it appears in the UI on load

* render History

updating the History Form and Code Behind to pull the history of changes for the connector and render them in the UI.

* Revert "History UI (#279)" (#280)

This reverts commit 550e70c7cc7ad5fb515ea7f763a1838da6a123c7.

* History UI (#281)

* load settings history

when the UI is loaded, history is retrieved and stored in memory

* History Form in Build

When the project is built, include the History form

* Form and Code Behind

History Form and History Form code behind placeholders. Will compile with build, but currently do not render the History in the UI.

* History Form in UI

Including the History Form as a Wizard Step so it appears in the UI on load

* render History

updating the History Form and Code Behind to pull the history of changes for the connector and render them in the UI.

* History reference

missing history reference in build

* Extending logging (#273)

* Get-SCSMWorkItemParent logging

Updating the Get-SCSMWorkItemParent function with New-SMExcoEvent functions instead of write-verbose

* Get-AzureEmailLanguage logging

Adding logging for when the language of the email is identified by Azure Cognitive Services

* Get-TierMembers logging

Adding logging for when the users of a Support Group (enum) are requested

* Get-AssignedToWorkItemVolume logging

Introducing logging when a user's active assigned work count is requested. Given the potential size of a support group, this requires verbose logging as it could potentially generate a lot of events.

* Readme and Inline Notes (#282)

* history form

screenshot for history form in the Settings UI

* Settings UI - History Form

screenshot and brief explanation of the History form in the UI

* version notes

updating notes to include relevant GitHub issues

* History Form and Logging (#283)

* null history handling

In the event a feature was never set (null) or set to null through PowerShell, the SDK call for history fails trying to parse the value to a string

* dynamic analyst logging

introducing logging for the Set-AssignedToPerSupportGroup function so there is clearer path as to how and why Get-TierMembers is used