user in memory only
adding error handling for when the user updating a work item does not exist in the CMDB and the configuration does not create them. In this event, the user object is created in memory alone so that subsequent actions can continue to occur. In the event of updating a relationship, try/catch blocks have been added. Logging placeholders have been created for New-SMExcoEvent