merge conflict on #292 (#295)
Re-doing the branch and associated changes. If the support group mapping cannot be obtained for any reason i.e. MP doesn't exist, the mapping doesn't exist, or AD can't be contacted then the function returns a $false result and logs an event. If the members of a Support Group can't be retrieved, return a null result and log an event. updating formatting/style of powershell