Managing Users not found in CMDB (#294)
* else condition for new users

if the user isn't found in SCSM and create users in CMDB is false. Then there needs to be a way to handle the Affected User in a New Work Item as well as handle who left the comment in an Update Work Item

* extending Create-UserInCMDB

creating the ability to create an SCSM User in memory only

* user in memory only

adding error handling for when the user updating a work item does not exist in the CMDB and the configuration does not create them. In this event, the user object is created in memory alone so that subsequent actions can continue to occur. In the event of updating a relationship, try/catch blocks have been added. Logging placeholders have been created for New-SMExcoEvent

* new work item, AU memory only

adding error handling for when the user creating a work item does not exist in the CMDB and the configuration does not create them. In this event, the user object is created in memory alone so that subsequent actions can continue to occur. In the event of updating a relationship, try/catch blocks have been added. Logging placeholders have been created for New-SMExcoEvent

* revert AML catch

this should not be part of this branch

* extended logging

adding logging when attempting to set relationships to user objects that only exist in memory. This would occur if Create Users in CMDB = false

* wrong source

while updating documentation discovered the wrong source was defined for said smexco events

* incorrect event numbering

further review of the wiki updates showed that the previous commit that changed the source from "new-workitem" to "update-workitem" did not take into account the shift in event id numbering.