Reactivation Logic (#303)
* psuedocode for ResCat and ResDesc

When using the reactivate keyword _or_ replying to a resolved Incident + using change status on reply. The resolution category and description should be nulled to follow suit with the console/portal.

* new function

Creating the Reactivate-WorkItem function to simplify how reactivating Incidents/Problems occurs wherein the Status changes, the Resolution (Category), Resolution Description, and Resolved Date becomes null

* Status Change on Reply logic

Changing the Status of an Incident based on who replies now takes into account an Incident that is currently Resolved. Thereby enforcing usage of the [reactivate] keyword with associated actions

* Status filtering

Adding a filter that prevents Resolved or Closed from appearing the Incident Status dropdown for changing status based on who replies. In doing so, this prevents a scenario wherein someone could choose Resolved/Closed and thereby skip processes associated with these statuses.

* and instead of or

The comparison used here should be an "and" and not an "or"