v3.3.1 (#296)
* declaring EWS dependency (#284)

Exchange Web Services is packaged and hosted on Nuget

* naming precision (#289)

While the chances of having similar naming conventions to these classes/management packs are slim given the context, they are never the less present. With the introduction of the "$" the possibility of returning multiple results is removed.

* merge conflict on #292 (#295)

Re-doing the branch and associated changes. If the support group mapping cannot be obtained for any reason i.e. MP doesn't exist, the mapping doesn't exist, or AD can't be contacted then the function returns a $false result and logs an event. If the members of a Support Group can't be retrieved, return a null result and log an event. updating formatting/style of powershell

* Managing Users not found in CMDB (#294)

* else condition for new users

if the user isn't found in SCSM and create users in CMDB is false. Then there needs to be a way to handle the Affected User in a New Work Item as well as handle who left the comment in an Update Work Item

* extending Create-UserInCMDB

creating the ability to create an SCSM User in memory only

* user in memory only

adding error handling for when the user updating a work item does not exist in the CMDB and the configuration does not create them. In this event, the user object is created in memory alone so that subsequent actions can continue to occur. In the event of updating a relationship, try/catch blocks have been added. Logging placeholders have been created for New-SMExcoEvent

* new work item, AU memory only

adding error handling for when the user creating a work item does not exist in the CMDB and the configuration does not create them. In this event, the user object is created in memory alone so that subsequent actions can continue to occur. In the event of updating a relationship, try/catch blocks have been added. Logging placeholders have been created for New-SMExcoEvent

* revert AML catch

this should not be part of this branch

* extended logging

adding logging when attempting to set relationships to user objects that only exist in memory. This would occur if Create Users in CMDB = false

* wrong source

while updating documentation discovered the wrong source was defined for said smexco events

* incorrect event numbering

further review of the wiki updates showed that the previous commit that changed the source from "new-workitem" to "update-workitem" did not take into account the shift in event id numbering.

* template error handling (#291)

Adding error handling for when Dynamic Analyst Assignment is used when the TierQueue/Support Group can't be determined on the Template. This addresses both missing enums on the template or an OOB template that is used/duplicated.

* Email and File Attachment logging (#293)

* add email event

if for whatever reason an email can't be attached to a work item, log an event. If an email exceeds the maximum allowed file size or attachment count, log an event

* try/catch block

adding try/catch block for file additions

* logging event

if a file attachment can't be added cite the sender, workitemid, and the exception in a new event

* formatting - 1

re-aligning indentation of curly braces

* formatting - 2

once again, aligning indentation/braces

* formatting - 3

further alignment changes

* formatting - 4

re-aligning indentation of new try block for Attach-FileToWorkItem

* Error handling for AML (#290)

* error handling for AML prediction

Addressing a scenario where AML makes a prediction about an enum that has since been removed from SCSM

* re-ordering events

incorrectly created new AML Events for the SR portion of New-WorkItem as opposed to use the same ones in an Incident

* logging level logic

updating logging events so they only occur as is configured in the settings

* logging logic for Support Group functions (#297)

* logging logic

adding logging level condition before writing an event

* function call syntax

just modifying the order of parameter/value pairs

* ACS priority logging (#299)

* ACS priority logging

Adding logging for when ACS boundaries are retrieved from the settings mp

* event log sources

Defining ACS Work Item Priority as a Logging Source

* param order (#300)

changing parameter/value order to be consistent throughout the connector

* function optimizations (#302)

* if to switch

simplifying SCOM DA retrieval

* simplifying switch

Instead of performing a switch against the work item type, perform a switch against the message type.

* if to switch

compressing Dynamic Analyst Assignment down by moving to a switch statement

* initial logic check (#304)

Caught during review. Lack of condition for when the user is found would result in a matched user always being created in memory

* Reactivation Logic (#303)

* psuedocode for ResCat and ResDesc

When using the reactivate keyword _or_ replying to a resolved Incident + using change status on reply. The resolution category and description should be nulled to follow suit with the console/portal.

* new function

Creating the Reactivate-WorkItem function to simplify how reactivating Incidents/Problems occurs wherein the Status changes, the Resolution (Category), Resolution Description, and Resolved Date becomes null

* Status Change on Reply logic

Changing the Status of an Incident based on who replies now takes into account an Incident that is currently Resolved. Thereby enforcing usage of the [reactivate] keyword with associated actions

* Status filtering

Adding a filter that prevents Resolved or Closed from appearing the Incident Status dropdown for changing status based on who replies. In doing so, this prevents a scenario wherein someone could choose Resolved/Closed and thereby skip processes associated with these statuses.

* and instead of or

The comparison used here should be an "and" and not an "or"

* testing/reviewing dev (#305)

* more ACS error handling

In the event ACS scoring can't be returned a null value would be returned and the Set-SCSMObject would fail

* adding logging

adding new logging events if the score can't be written and/or written to an optionally configured class extension

* 365 username logging (#308)

Updating logging event per discussion in #307, correcting a minor spelling mistake with "successful", and removing variable in favor of placing the regex condition directly inside of the IF statement.

* inline notes (#306)

* inline notes

updating notes to reference pull requests addressed in upcoming release

* versioning info

adding version number into notes

* recent issue fix

Updating notes with further changes in the connector

* Change Status on Reply events (#309)

* error handling for setting enum

Ensures that if the Status enum can't be set based on who replied that the connector can continue to process when using workflows

* event logging

adding new event for Update-WorkItem if the Status can't be changed

* schedule-workitem parameter (#310)

The Work Item type string is no longer a needed parameter to use the function. Updating function and subsequent calls as well as an additional logging event should the Scheduled Start/End Times not be able to be updated.

* null comment (#311)

If the reply to a Work Item cannot identify the location of "from:" and the body is null, then Add-ToActionLogEntry would throw an error about a null value

* version changes (#312)

adding another item to the version notes

* logging level logic (#313)

caught during review. adding missing logging level logic.