optimization and error handling
Performing checks to ensure the subject and/or body contain values before continuing through to the function. Simplifying the variable that is tested/eventually used, down to $searchQuery. Finally writing an empty string in the event one or both search results return no values. In this way, New-WorkItem correctly parses HTML results and subsequently chooses the correct template.