v3.4 (#323)
* File Attachment Support for Activities (#320)

* Update smletsExchangeConnector.ps1

Add logic to attach file to parent work item when the work item is a manual or review activity.

* Work Item guid

Updating Work Item guid property

* powershell case

updating formatting for smlets

* Work Item ID

Attach-FileToWorkItem takes the ID (i.e. CR1234) of a Work Item.

* moving logging call

Moving the logging from Update-WorkItem to Attach-FileToWorkItem. This ensures every file attached is logged and creates the ability to provide more detailed logging around the file itself.

Co-authored-by: Adam <adhocadam@protonmail.com>

* readme: File Attachments for Activities (#322)

* feature icon

created a new icon for the readme to highlight an incoming feature on #320

* Activity File Attachments

updating readme to include notes on said feature

* column width

resizing columns

* wf: title/subject error handling (#315)

possibility of subject being longer than 200 characters

* supported endpoint (#321)

Per Cireson's API documentation, the GetHTMLArticlesFullTextSearch endpoint is deprecated. The new FullTextSearch endpoint returns the same values in testing and does not require a Cireson Portal User object to perform a search.

* trimming search query (#318)

ensures a clean split into an array of all of the words that were in the subject/body

* suggestion url optimization and error handling (#325)

Performing checks to ensure the subject and/or body contain values before continuing through to the function. Simplifying the variable that is tested/eventually used, down to $searchQuery. Finally writing an empty string in the event one or both search results return no values. In this way, New-WorkItem correctly parses HTML results and subsequently chooses the correct template.

* obtaining a reviewer (#327)

When obtaining a User from a Reviewer, it's possible the User was not defined in SCSM. Which means when the connector runs against an RA, an error would be thrown to the effect of "Exception Message: Id='' -- Unrecognized Guid format." because Get-SCSMObject would fail to parse a null value.

Now the related object's existence is confirmed before moving forward with the processing.

* notes and readme for 3.4 (#326)

* local icon

noticed the icon link is broken. Going to update the readme with a local repo image

* updating icon url

changing link to a local repo URL

* inline notes

updating the inline notes on the connector's powershell

Co-authored-by: Brad Zima <infiniterecursion@users.noreply.github.com>