Support for Custom Rules (Management Pack) (#337)
* Message Class enum and notes

Creating a list of known message classes the connector supports out of the box. Further breaking sections out with inline notes to highlight their purpose

* Custom Rule class

Adding a new admin class (SMLets.Exchange.Connector.AdminSettings.CustomRule) that allows the creation of one or many "Custom Rules" that defines a friendly name, message class per SMLets.Exchange.Connector.MessageClassEnum, message part, the regex pattern, what to create/update, and finally the property on the class the matched property should be written to

* custom rules wizard pane

Adding a new window (wizard pane) for creating, updating, and deleting Custom Rules. Also updating build process so the page is included and updating validation rules to include empty string checks for the CustomRules.xaml page

* readme, custom rules icon

uploading icon for custom rules

* readme placeholder

adding section for custom rules

* v4 feature typo

should read "Custom Rules" and not "External Ticketing"

* placeholder text update

adding some use case for the feature along with a link to the wiki.

* New Custom Event, Invoke-CustomRuleAction

When using Custom Rules, the defined Custom Rule Item Type to create/update could be overridden to a value that is not provided out of box in the Settings UI. In this case, a truly custom action is being performed.

* custom rule icon

adding an image reference/icon for the Custom Rule Class