refactor, Update-WorkItem + Confirm-WorkItem
Now that Add-FileToSCSMObject takes an SMObject parameter instead of a WorkItemID string, Update-WorkItem's parameters must be adjusted to receive an SMObject instead of a WorkItemID string so the call to Add-FileToSCSMObject succeeds.

Since the core processing loop of the connector that could invoke Update-WorkItem already retrieves the SMObject, that is now passed instead of the WorkItemID string. In doing so, duplicate calls to retrieve the Work Item to update within SCSM are now eliminated.

Logging within the Update-WorkItem function has been adjusted to account for this change.