Support for Custom Rules (PowerShell) (#340)
* function, test-emailpattern

creating placeholder for function/draft the pull request

* refactor, Add-EmailToWorkItem

Instead of using the vaguely named $fileAttachmentRelClass variable for "System.WorkItemHasFileAttachment", use $wiHasFileAttachRelClass

* refactor, Add-FileToSCSMObject

function is being updated to support Config Items in addition to Work Items. Logging change to cite the Name of the Object instead of the Work Item ID. This is a breaking change that will require refactoring of other functions throughout the connector so they provide an SMObject instead of a WorkItemID

* refactor, Update-WorkItem + Confirm-WorkItem

Now that Add-FileToSCSMObject takes an SMObject parameter instead of a WorkItemID string, Update-WorkItem's parameters must be adjusted to receive an SMObject instead of a WorkItemID string so the call to Add-FileToSCSMObject succeeds.

Since the core processing loop of the connector that could invoke Update-WorkItem already retrieves the SMObject, that is now passed instead of the WorkItemID string. In doing so, duplicate calls to retrieve the Work Item to update within SCSM are now eliminated.

Logging within the Update-WorkItem function has been adjusted to account for this change.

* can't create Problem

New-WorkItem does not create a Problem despite logging it

* function, Test-EmailPattern

introduction of Test-EmailPattern that evaluates custom regular expressions found in an email that are NOT the out of box patterns per configuration within the Settings UI

* reference variables

Adding inline documentation to the Invoke-CustomRuleAction function so it's clear what variables are passed in, how they are used, and how this function can managed given any number of Custom Rules (IF/ELSE or SWITCH)